services/brain/requirements.txt
redis
python-dotenv
openai
